+++ Sources/VComponents/Components/Containers/Disclosure Group/VDisclosureGroupUIModel.swift
@@ -26,7 +26,7 @@ public struct VDisclosureGroupUIModel {
     }
 
     // MARK: Properties - Corners
-    /// Corner radii . Set to to `15`s.
+    /// Corner radii. Set to to `15`s.
     public var cornerRadii: RectangleCornerRadii = .init(15)
 
     /// Indicates if horizontal corners should switch to support RTL languages. Set to `true`.