+++ Sources/VComponents/Components/Modals/SideBar/VSideBarExtension.swift
@@ -9,7 +9,6 @@ import SwiftUI
 import VCore
 
 // MARK: - Bool
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)
@@ -100,7 +99,6 @@ extension View {
 }
 
 // MARK: - Item
-@available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 @available(visionOS, unavailable)